refactor costing logic into service
