--- Datahub.Graph.Functions/CostManagement/CostReturnRecord.cs
@@ -1,7 +1,0 @@
-namespace Datahub.Graph.Functions.CostManagement;
-
-public record CostReturnRecord
-{
-    
-    public int NumberOfProjectsUpdated { get; init; }
-};
